Refactor resources/dokku; add basic caching
This is a big commit touching the entire codebase. I needed to figure
out what sort of structure I was going to stick with and it was just
simpler to vomit it all out in one big chunk until it got to the point
where I knew I could make smaller, incremental changes.