Add cache control via socket
Tracking the dokku event log is error prone and doesn't cover all
cases (e.g. stopping an app - there is no event logged for this).
Therefore, if caching is to function properly a dokku plugin will be
necessary. Dokku exposes all necessary events (even for events such as
stopping an app) to plugins via hooks.

dokku-api now creates two services bound to two sockets: One for the
API, now located at $socket-dir/api/api.sock, which is meant to be
shared via a docker mount with an app container; and another one,
located at $socket-dir/plugin.sock, for cache invalidation and
prefetching. This second one is meant to be used by the (planned)
dokku-api plugin for Dokku, which will subscribe to the necessary event
hooks and issue cache control commands to dokku-api based on the content
and source of those events.